+++ R/draw_map_focus_one_project.R
@@ -9,12 +9,9 @@
 #' @param zoom_level An integer. The zoom level for the map. Defines
 #' default zoom level.
 #' 
-#' @importFrom leaflet addCircleMarkers addPolygons
-#' @importFrom dplyr filter select mutate rename
-#' @importFrom tibble as_tibble
-#' @importFrom tidyr separate_rows
-#' @importFrom sf st_drop_geometry
-#' 
+#' @importFrom leaflet addCircleMarkers addPolygons addPolylines labelOptions
+#' @importFrom dplyr filter
+#'  
 #' @return A leaflet object.
 #' 
 #' @export
@@ -25,12 +22,6 @@
 #' draw_map_focus_one_project(
 #'   projects_data_sf = toy_projects_data_sf,
 #'   id_project = "1+1=3  PGV03.038", 
-#'   cantons_sf = toy_cantons_sf
-#' )
-#'
-#' draw_map_focus_one_project(
-#'   projects_data_sf = toy_projects_data_sf,
-#'   id_project = "FM_ProPCC+", 
 #'   cantons_sf = toy_cantons_sf
 #' )
 draw_map_focus_one_project <- function(
@@ -44,81 +35,78 @@ draw_map_focus_one_project <- function(
     cantons_sf <- read_cantons_sf()
   }
   
-  # Create a coord_sf_project object
-  projects_data_sf_filtered <- projects_data_sf |> 
-    filter(short_title == id_project) 
-  
-  coord_sf_project <- projects_data_sf_filtered |> 
-    select(short_title, geometry)
-  
-  # Extract target cantons from project_data_sf
-  project_data_cantons_sf <- projects_data_sf_filtered |> 
-    st_drop_geometry() |> 
-    select(short_title, geo_range_id) |> 
-    separate_rows(
-      geo_range_id, 
-      sep = ", "
-    )
-  
-  # Create a cantons_sf_project object with column target_cantons (TRUE/FALSE)
-  cantons_sf_project <- left_join(
-    x = cantons_sf |> 
-      select(HASC_1, NAME_1, geometry),
-    y = project_data_cantons_sf |> 
-      rename(HASC_1 = geo_range_id), 
-    by = "HASC_1"
-  ) |> 
-    mutate(
-      target_cantons = ifelse(
-        is.na(short_title),
-        FALSE, 
-        TRUE
-      )
-    ) |> 
-    select(- short_title)
+  # Extract the elements needed for the map
+  ## The coordinates of the project
+  ## The polygons of the canton with the info influenced/not influenced
+  ## If appropriate, the lines between the projects and the cantons influenced
+  geo_elements_influence <- get_influence_project(
+    projects_data_sf = projects_data_sf,
+    id_project = id_project,
+    cantons_sf = cantons_sf
+  )
   
   # Draw the map
-  # Display target cantons in psch orange / others psch blue
-  # Display selected project in psch orange
-  # Nice to have
-  # Display lines radiating from project to centroid of target cantons
-  map_with_not_influenced_cantons <- draw_map_base(zoom_level) |> 
+  ## Add the cantons not influenced
+  map_with_cantons_not_influenced <- draw_map_base(zoom_level) |> 
     addPolygons(
-      data = cantons_sf_project |> 
-        filter(target_cantons == FALSE),
+      data = geo_elements_influence$cantons_sf_project |> 
+        filter(
+          target_cantons == FALSE
+          ),
       weight = 1,
       color = psch_blue(),
       fillColor = psch_blue(),
-      fillOpacity = 0.1
+      fillOpacity = 0.2
     ) 
-  
-  if (nrow(cantons_sf_project |> 
-        filter(target_cantons == TRUE)) > 0) {
+   
+  ## Add the cantons influenced if appropriate and the lines between the project and them
+  if (!is.null(geo_elements_influence$cantons_influenced_lines)) {
     
-    map_with_influenced_cantons <- map_with_not_influenced_cantons |> 
+    # Print the cantons in orange
+    map_with_cantons_influenced <- map_with_cantons_not_influenced |> 
       addPolygons(
-      data = cantons_sf_project |> 
-        filter(target_cantons == TRUE),
+      data = geo_elements_influence$cantons_sf_project |> 
+        filter(
+          target_cantons == TRUE
+        ),
       weight = 1,
       color = psch_orange(),
       fillColor = psch_orange(),
       fillOpacity = 0.5,
-      label = ~ as.character(NAME_1)
-    ) 
+      label = ~ gsub("^CH\\.", "", as.character(HASC_1)), 
+      labelOptions = labelOptions(
+        style = list("font-size" = "12px")  # Adjust the font size here
+        )
+      ) 
     
+    # Print the lines in orange
+    map_geo_influence <- map_with_cantons_influenced |> 
+      addPolylines(
+       data = geo_elements_influence$cantons_influenced_lines, 
+       color = psch_orange(), 
+       weight = 2
+      )
+      
   } else {
     
-    map_with_influenced_cantons <- map_with_not_influenced_cantons
+    map_geo_influence <- map_with_cantons_not_influenced
     
   }
     
-  map_with_influenced_cantons |>
+  map_project_and_geo_influence <- map_geo_influence |>
     addCircleMarkers(
-      data = coord_sf_project,
+      data = geo_elements_influence$coord_sf_project,
       color = psch_orange(),
       stroke = FALSE,
       fillOpacity = 0.8,
       label = ~ as.character(short_title)
     )
+  
+  class(map_project_and_geo_influence) <- c(
+    class(map_project_and_geo_influence),
+    "one-project"
+  )
+  
+  return(map_project_and_geo_influence)
   
 }