+++ R/mod_projects_selection.R
@@ -10,21 +10,8 @@
 mod_projects_selection_ui <- function(id){
   ns <- NS(id)
   tagList(
-    fluidRow(
-      column(9),
-      column(
-        3,
-        tags$div(
-          materialSwitch(
-            inputId = "language_switch",
-            label = "DE",
-            inline = TRUE
-          ),
-          tags$span("FR")
-        )
-      )
-    ),
-    fluidRow(
+
+    div(
       h2("Hallo!") |>
         with_i18n("app-title"),
       checkboxGroupInput(